+++ Atelier Sophie DX/CHANGELOG.md
@@ -1,4 +1,7 @@
 # Change log
+
+## 2024/10/22
+1. added scripts to change screen mode from "window mode" to "borderless fullscreen"
  
 ## 2022/07/16
  Migrate from https://opencheattables.org